requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
ftfy==6.2.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
gitignore_parser==0.1.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
grpcio==1.67.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.0
idna==3.7
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
json5==0.9.25
jsonpickle==3.2.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmcv==2.2.0
mmengine==0.10.5
model-index==0.1.11
mpmath==1.3.0
munch==4.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-autoremove==0.10.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.28.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.19.1
rsa==4.9
sacred==0.8.5
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seml==0.5.1
Send2Trash==1.8.3
sentry-sdk==2.11.0
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
submitit==1.5.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.11
tinycss2==1.3.0
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.65.2
traitlets==5.14.3
triton==3.0.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.20
wandb==0.17.5
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
zipp==3.19.2
